Insert Many Performance Tests
